# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Capability": {
        "description": "What Vokara needs from a provider, configured independently (ADR-011).",
        "enum": [
          "generation",
          "embeddings"
        ],
        "title": "Capability",
        "type": "string"
      },
...
      "DisclosureModel": {
        "description": "The full text of the disclosure of art. V, to be shown on screen.",
        "properties": {
          "acknowledged": {
            "title": "Acknowledged",
            "type": "boolean"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "acknowledged_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Versión efectivamente acusada; un acuse viejo no cubre un texto nuevo.",
            "title": "Acknowledged Version"
          },
          "body_md": {
            "description": "Texto completo en español. Nunca solo un enlace.",
            "title": "Body Md",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "body_md",
          "acknowledged"
        ],
        "title": "DisclosureModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API local de Vokara. Sin autenticación (ADR-008): la instancia sirve a una sola persona, la que la instaló, y escucha únicamente en la interfaz local.",
    "title": "Vokara API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/setup/disclosure": {
      "get": {
        "description": "The complete text: art. V forbids it being only a link or only the README.",
        "operationId": "get_disclosure_api_v1_setup_disclosure_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisclosureModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          }
        },
        "summary": "Texto vigente de la divulgación y estado de su acuse",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/setup/disclosure-acknowledgement": {
      "post": {
        "description": "Recorded with its timestamp and the version accepted (FR-002, R-29).",
        "operationId": "post_disclosure_acknowledgement_api_v1_setup_disclosure_acknowledgement_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisclosureAcknowledgementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStateModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "DISCLOSURE_ACKNOWLEDGEMENT_REQUIRED · DISCLOSURE_ALREADY_ACKNOWLEDGED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          }
        },
        "summary": "Registra el acuse explícito de la divulgación",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/setup/email": {
      "get": {
        "description": "Carries the disclosure with it, because FR-012 requires it **before**.\n\nA warning that arrives after the form has been filled in is not a warning.",
        "operationId": "get_email_step_api_v1_setup_email_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailStepModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          }
        },
        "summary": "Estado del paso de correo",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/setup/email/link": {
      "post": {
        "description": "Verifies the label exists before taking the link as established (FR-013).\n\nIts three failures leave with their own code and none of them blocks\nanything: the step stays skippable at every point.",
        "operationId": "post_email_link_api_v1_setup_email_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailStepModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "EMAIL_APP_PASSWORD_REJECTED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "EMAIL_LABEL_NOT_FOUND · VALIDATION_ERROR"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "EMAIL_PROVIDER_UNREACHABLE"
          }
        },
        "summary": "Vincula la cuenta de correo y verifica la etiqueta designada",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/setup/email/skip": {
      "post": {
        "description": "One action, and the first run can conclude. A valid ending (FR-011).",
        "operationId": "post_email_skip_api_v1_setup_email_skip_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStateModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          }
        },
        "summary": "Omite el paso de correo",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/setup/providers/catalog": {
      "get": {
        "description": "The closed list, already resolved: the frontend renders, it does not branch.\n\nThe cost travels here so it can be shown **before** any key is asked for\n(FR-005), estimated separately for each capability.",
        "operationId": "get_provider_catalog_api_v1_setup_providers_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCatalogModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          }
        },
        "summary": "Proveedores ofrecibles por capacidad, con su costo estimado",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/setup/providers/{capability}": {
      "get": {
        "description": "`null` when nothing is configured. Never the credential (FR-008).",
        "operationId": "get_provider_configuration_api_v1_setup_providers__capability__get",
        "parameters": [
          {
            "in": "path",
            "name": "capability",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Capability"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ProviderConfigurationModel"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Provider Configuration Api V1 Setup Providers  Capability  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          }
        },
        "summary": "Configuración vigente de una capacidad",
        "tags": [
          "setup"
        ]
      },
      "put": {
        "description": "The preflight runs **here**, at save time, never deferred (FR-006).\n\nThe three results that do not allow progress leave through the error\ncatalogue with their own status and code, because they are three different\nsituations and the frontend has to be able to tell them apart without\nreading prose (contracts/errors.md).",
        "operationId": "put_provider_configuration_api_v1_setup_providers__capability__put",
        "parameters": [
          {
            "in": "path",
            "name": "capability",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Capability"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderConfigurationModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "MODEL_NOT_AVAILABLE · PROVIDER_CREDENTIAL_REJECTED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "PROVIDER_NOT_OFFERED · VALIDATION_ERROR"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "PROVIDER_QUOTA_EXCEEDED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "PROVIDER_UNREACHABLE"
          }
        },
        "summary": "Configura una capacidad y ejecuta su preflight",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/setup/providers/{capability}/degradation-acknowledgement": {
      "post": {
        "description": "The only way a capability without guarantee becomes usable (FR-007.3).\n\nAnswered 409 when the current preflight is not `capability_unverified`:\nthere is no degradation to acknowledge, and recording one would be a\nconsent to nothing.",
        "operationId": "post_degradation_acknowledgement_api_v1_setup_providers__capability__degradation_acknowledgement_post",
        "parameters": [
          {
            "in": "path",
            "name": "capability",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Capability"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderConfigurationModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "DEGRADATION_ACKNOWLEDGEMENT_REQUIRED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "INTERNAL_ERROR"
          }
        },
        "summary": "Acuse específico de una degradación explícita",
        "tags": [
          "setup"
        ]
      }
    },
...
  }
}